Improve boundary processing in shortbread topic
More consistent processing of boundaries in shortbread_v1. Deciding when
something is a boundary is now done only in one place. And the code is
more robust, checking that the way info is there before using it.

(This is now the same as in the shortbread_v1_gen code which already had
those changes.)

See #2