Allow tilekiln generation to specify metadata
Tilekiln configs can include optional data for the creation of
a TileJSON document. All of the metadata is optional except for
id, which is used for storage and construction of URLs. In the
themepark config generation we set a default id, but otherwise
it takes it from the options passed.