shortbread: boundaries: Require admin_level to be one of the valid values
Shortbread says 2 and 4 are the valid values of admin_level, so this
doesn't look at a relation that is only tagged with
type=boundary boundary=administrative and has no admin_level tagging.

This doesn't strictly comply with shortbread because a member of a
admin_level 3 and admin_level 4 relation should get admin_level set
to 3 but I don't believe that's the intent of the spec.