shortbread: boundaries: Fix disputed boundary processing
Previously a member of multiple relations would take its disputed status
from the last relation processed. For example, processing a way that is
a member of two relations, with

    type=boundary boundary=administrative admin_level=2
    type=boundary boundary=disputed admin_level=2

would lead to order-dependent results, where disputed only comes from
the last relation processed. Another issue was that a relation with

    type=boundary boundary=disputed

and no other tags would not be considered as tonumber(tags.admin_level)
is not truth-like.

Instead, what is done is to get the disputed status for the relation,
and then set it to true for the rinfos in that relation, not setting
any existing rinfos to false. They are only set false when creating the
entry in rinfos for the first time.