Fix shortbread boundary processing in regards to disputed boundaries
The way disputed boundaries are implemented was not correct. Generally
boundaries should be processed as if disputed boundaries do not exists.
But then boundary lines are flagged as disputed if they are created
from ways

* that have the tag disputed=yes, or
* that are in a relation tagged boundary=disputed with no admin_level
  set or an admin_level smaller or equal to the admin_level that
  the boundary line has (which is the smallest of any boundary relation
  the way is a member of)

Note that the relation that marks a way as disputed is not the same
relation that marks a way as part of some specific admin boundary. The
first is tagged boundary=disputed, the second is tagged
boundary=administrative.

This is based on the work in #10 but rewrites the code to be (hopefully)
easier to understand.

Note that this only fixes the version for shortbread_v1, the version for
shortbread_v1_gen also needs updating which will come later.