New env var THEMEPARK_PATH can be used to set theme search path
As usual use a colon (:) to seperate items in the path.

Initial search path is now whetever is in THEMEPARK_PATH and then the
"themes" directory of the repo.

The themepark:add_theme_dir() function now prepends the specified
directory to the search path instead of appending it.

In debug mode the search path is printed at the start and whenever it
changes.