Add new "explore" theme with some initial topics
OpenStreetMap contains an insane amount of complex data. If you want to
use some of that data a first step is often to explore what's there,
filter out the stuff that looks interesting, creating some experimental
maps, figuring out what's useable and how. The topics in this theme are
meant as starting points for such an exploration. They are intended to
be used with a tool such as QGIS to quickly display what's there in
different ways.

Initial topics are "coastline", "postcode", and "restrictions".